+++ src/utils.py
@@ -46,6 +46,8 @@ Functions in this module
 
 * *CodonMutsCumulFracs* : cumulative counts of codon mutations.
 
+* *ParseNSMutFreqBySite* : Parse nonsynonymous mutation frequencies at each site from a DMS counts file.
+
 Function documentation
 ---------------------------
 
@@ -1217,6 +1219,38 @@ def CodonMutsCumulFracs(codon_counts):
     multi_nt_all_cumulfracs = [multi_nt_all_cumulfracs[n] for n in range(maxkey + 1)]
     multi_nt_syn_cumulfracs = [multi_nt_syn_cumulfracs[n] for n in range(maxkey + 1)]
     return (all_cumulfracs, all_counts, syn_cumulfracs, syn_counts, multi_nt_all_cumulfracs, multi_nt_all_counts, multi_nt_syn_cumulfracs, multi_nt_syn_counts)
+
+def ParseNSMutFreqBySite(countsfile, chartype):
+    """Parses a deep mutational scanning counts file and returns
+    the nonsynonymous mutation frequency at each site.
+    
+    *countsfile* is a string that gives the name of a file.
+    
+    *chartype* is the type of character, with valid values consisting of:
+
+        - *codon* : DNA codons.
+
+        - *aminoacids_nostop* : amino acids not including stop codons.
+
+        - *aminoacids_withstop* : amino acids including stop codons (``*``).
+    
+    The returned item is a list of tuples of (site, nonsynonymous frequency).
+    
+    """
+    if chartype.upper() == 'CODON':
+        translate_to_aa = True
+    elif chartype.upper() == 'AMINOACIDS_NOSTOP' or chartype.upper() == 'AMINOACIDS_WITHSTOP':
+        translate_to_aa = False
+    countsdict = dms_tools.file_io.ReadDMSCounts(countsfile, chartype, translate_codon_to_aa = translate_to_aa)
+    convert = lambda text: int(text) if text.isdigit() else text.lower() 
+    alphanum_key = lambda key: [ convert(c) for c in re.split('([0-9]+)', key) ]
+    sites = sorted(countsdict.keys(), key = alphanum_key)
+    tuplist = []
+    for site in sites:
+        wt_aa = countsdict[site]['WT']
+        ns_mut_freq = 1 - countsdict[site]['F_%s' % wt_aa]
+        tuplist.append((site,ns_mut_freq))
+    return tuplist
 
 
 # Test with doctest