+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Changelog
 
 * Added ``translate_codon_to_aa`` and ``return_as_df`` options to dms_tools.file_io.ReadDMSCounts().
 
+* Added ``dms_tools.utils.ParseNSMutFreqBySite()``.
+
 1.1.16
 --------
 * Added ``--mincounts`` option to ``dms_diffselection``, and make ``dms_logoplot`` properly plot differential selection with missing values.